refactor: rename trace_route.rs to file_tracker.rs
