+++ rook/src/git/mod.rs
@@ -1,9 +1,9 @@
+mod file_tracker;
 mod link_extractor;
 mod repo;
-mod trace_route;
 mod url;
 
+pub use file_tracker::*;
 pub use link_extractor::*;
 pub use repo::*;
-pub use trace_route::*;
 pub use url::*;
